move Hellen-specific stuff to hellen_common.cpp
