+++ firmware/config/engines/custom_engine.h
@@ -36,8 +36,3 @@ void setBoschHDEV_5_injectors();
 void setHellen72etb();
 void setRotary();
 void setVrThresholdTest();
-
-void setHellen144LedPins();
-void setHellen176LedPins();
-
-void detectBoardType();
